+++ tests/test_score.py
@@ -1,7 +1,6 @@
 import unittest
 
-from peptacular.fragment import Fragment
-from peptacular.score import get_fragment_matches, FragmentMatch, get_matched_indices
+import peptacular as pt
 
 
 class TestScore(unittest.TestCase):
@@ -10,34 +9,34 @@ class TestScore(unittest.TestCase):
         # Test basic functionality with 'th' tolerance
         mz_spectrum1 = [100, 200, 300]
         mz_spectrum2 = [99.9, 100.1, 200, 200.05, 300.1, 300.2]
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 0.2, 'th')
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 0.2, 'th')
         assert result == [(0, 2), (2, 4), (4, 6)]
 
         # Test basic functionality with 'th' tolerance
         mz_spectrum1 = [100, 200, 300]
         mz_spectrum2 = [99.9, 100.1, 200, 200.05, 300.1]
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 0.2, 'th')
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 0.2, 'th')
         assert result == [(0, 2), (2, 4), (4, 5)]
 
         # Test with 'ppm' tolerance
         mz_spectrum1 = [100, 200, 300]
         mz_spectrum2 = [99.9, 100.1, 200, 200.05, 300.1, 300.2]
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 1000, 'ppm')  # 0.1 ppm tolerance
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 1000, 'ppm')  # 0.1 ppm tolerance
         assert result == [(0, 2), (2, 4), (4, 6)]
 
         # Test with no matches
         mz_spectrum1 = [50, 150, 250]
         mz_spectrum2 = [100, 200, 300]
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 0.05, 'th')
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 0.05, 'th')
         assert result == [None, None, None]
 
         # Test with empty spectra
         mz_spectrum1 = []
         mz_spectrum2 = [100, 200, 300]
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 0.1, 'th')
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 0.1, 'th')
         assert result == []
 
         mz_spectrum1 = [100, 200, 300]
         mz_spectrum2 = []
-        result = get_matched_indices(mz_spectrum1, mz_spectrum2, 0.1, 'th')
+        result = pt.get_matched_indices(mz_spectrum1, mz_spectrum2, 0.1, 'th')
         assert result == [None, None, None]